Commit of working Irish Covid data dashboard
This is the first version of the data dashboard which, using data from
gov.ie, displays information about the spread of Covid-19 in Ireland. It
is a recreation of the data dashboard built by the HSE rebuilt using
Plotly Dash for python. As well as recreating some of the basic plots
this also expands on some of the information provided by the HSE
dashboard such as using a choropleth map and including a 3 day rolling
average of new cases.