Commit of Project Titanic data science project
This is the first version of my analysis of the "Titanic: Machine
Learning from Disaster" dataset from Kaggle. In this commit there are
three jupyter notebooks:
- Titanic_Data_Exploration - An exploration of the data set and the
  relationships between certain data attributes.
- Titanic_Classification - An exploration of multiple classification
  models trained on the data with the best one being chosen to submit to
  the competition.
- Titanic_ANN - An exploration of the hyperparameters of an artificial
  neural network trained on the data with the best parameters chosen to
  submit to the competition.

The data from Kaggle and the results of the classification model and the
neural network have also been added. So far the best performing
classification model was the RandomForestClassifier with an accuracy
score of 0.76315 on the final competition data. The neural network
preformed better with a high score of 0.78947 placing the submission at
2700th out of ~20,000 submissions at time of writing.